refactor: recommend starship instead of powerline
The startship statusline is a lot more beginner friendly and involves a more straightforward installation. This can be
helpful for interns so they can "set it and forget it". Also starship has some handy interactive commands that helps them understand their statusline, a very valuable feature. And lastly, its written in rust, perfect :)